Move Exercise class to backend
